+++ lib/run/workers/ParticleJobs.h
@@ -54,6 +54,32 @@ public:
     virtual void evaluate(const RunSettings& global, IRunCallbacks& callbacks) override;
 };
 
+class MultiJoinParticlesJob : public IParticleJob {
+private:
+    int slotCnt = 3;
+    bool moveToCom = false;
+
+public:
+    MultiJoinParticlesJob(const std::string& name)
+        : IParticleJob(name) {}
+
+    virtual std::string className() const override {
+        return "multi join";
+    }
+
+    virtual UnorderedMap<std::string, JobType> getSlots() const override {
+        UnorderedMap<std::string, JobType> map;
+        for (int i = 0; i < slotCnt; ++i) {
+            map.insert("particles " + std::to_string(i + 1), JobType::PARTICLES);
+        }
+        return map;
+    }
+
+    virtual VirtualSettings getSettings() override;
+
+    virtual void evaluate(const RunSettings& global, IRunCallbacks& callbacks) override;
+};
+
 class TransformParticlesJob : public IParticleJob {
 private:
     struct {
@@ -342,20 +368,20 @@ public:
     virtual void evaluate(const RunSettings& global, IRunCallbacks& UNUSED(callbacks)) override;
 };
 
-class AnalysisJob : public IParticleJob {
-private:
-    Path outputPath = Path("report.txt");
-
+class CompareJob : public IParticleJob {
 public:
-    explicit AnalysisJob(const std::string& name)
+    explicit CompareJob(const std::string& name)
         : IParticleJob(name) {}
 
     virtual std::string className() const override {
-        return "analysis";
+        return "compare";
     }
 
     virtual UnorderedMap<std::string, JobType> getSlots() const override {
-        return { { "particles", JobType::PARTICLES } };
+        return {
+            { "test particles", JobType::PARTICLES },
+            { "reference particles", JobType::PARTICLES },
+        };
     }
 
     virtual VirtualSettings getSettings() override;